+++ lib/action_reporter/current.rb
@@ -1,3 +1,4 @@
+require "active_support/isolated_execution_state"
 require_relative "error"
 
 module ActionReporter