+++ spec/spec_helper.rb
@@ -23,4 +23,13 @@ Dir[File.expand_path("support/**/*.rb", __dir__)].each { |f| require_relative f 
 RSpec.configure do |config|
 end
 require "polyrun/rspec"
+Polyrun::RSpec.install_sharded_formatter_compat!
 Polyrun::RSpec.install_failure_fragments!
+Polyrun::RSpec.install_worker_ping!
+Polyrun::RSpec.install_example_debug!
+Polyrun::RSpec.install_example_rails_logging!
+Polyrun::RSpec.install_example_timeout!
+Polyrun::RSpec.install_example_prosopite!
+if %w[1 true yes].include?(ENV["POLYRUN_SPEC_QUALITY"]&.to_s&.downcase)
+  Polyrun::RSpec.install_spec_quality!
+end